fix: Remove GET /stages
