+++ plugins/pipelines/listStages.js
@@ -4,7 +4,8 @@ const boom = require('@hapi/boom');
 const joi = require('joi');
 const schema = require('screwdriver-data-schema');
 const pipelineIdSchema = schema.models.pipeline.base.extract('id');
-const stageListSchema = joi.array().items(schema.models.stage.base).label('List of stages');
+const nameSchema = schema.models.stage.base.extract('name');
+const stageListSchema = schema.models.stage.list;
 
 module.exports = () => ({
     method: 'GET',
@@ -20,6 +21,7 @@ module.exports = () => ({
 
         handler: async (request, h) => {
             const { pipelineFactory, stageFactory } = request.server.app;
+            const { name, sort, sortBy, page, count } = request.query;
             const pipelineId = request.params.id;
 
             return pipelineFactory
@@ -30,8 +32,30 @@ module.exports = () => ({
                     }
 
                     const config = {
-                        params: { pipelineId }
+                        params: { pipelineId },
+                        sort
                     };
+
+                    if (name) {
+                        config.params = {
+                            ...config.params,
+                            name
+                        };
+                    }
+
+                    if (sortBy) {
+                        config.sortBy = sortBy;
+                    }
+
+                    if (page) {
+                        config.paginate = config.paginate || {};
+                        config.paginate.page = page;
+                    }
+
+                    if (count) {
+                        config.paginate = config.paginate || {};
+                        config.paginate.count = count;
+                    }
 
                     return stageFactory.list(config);
                 })
@@ -49,6 +73,7 @@ module.exports = () => ({
             }),
             query: schema.api.pagination.concat(
                 joi.object({
+                    name: nameSchema,
                     search: joi.forbidden() // we don't support search for Pipeline list stages
                 })
             )
