+++ test/plugins/stages.test.js
@@ -5,7 +5,6 @@ const sinon = require('sinon');
 const hapi = require('@hapi/hapi');
 const hoek = require('@hapi/hoek');
 const testStage = require('./data/stage.json');
-const testStages = require('./data/stages.json');
 const testStageBuilds = require('./data/stageBuilds.json');
 
 sinon.assert.expose(assert, { prefix: '' });
@@ -79,140 +78,6 @@ describe('stage plugin test', () => {
 
     it('registers the plugin', () => {
         assert.isOk(server.registrations.stages);
-    });
-
-    describe('GET /stages', () => {
-        let options;
-
-        beforeEach(() => {
-            options = {
-                method: 'GET',
-                url: '/stages?page=1&count=3',
-                auth: {
-                    credentials: {
-                        username: 'foo',
-                        scmContext: 'github:github.com',
-                        scope: ['user']
-                    },
-                    strategy: ['token']
-                }
-            };
-        });
-
-        it('returns 200 and all stages', () => {
-            stageFactoryMock.list
-                .withArgs({
-                    paginate: {
-                        page: 1,
-                        count: 3
-                    },
-                    sort: 'descending'
-                })
-                .resolves(getStageMocks(testStages));
-
-            return server.inject(options).then(reply => {
-                assert.equal(reply.statusCode, 200);
-                assert.deepEqual(reply.result, testStages);
-            });
-        });
-
-        it('returns 200 and all stages with pipelineId', () => {
-            options.url = '/stages?pipelineId=123';
-            stageFactoryMock.list
-                .withArgs({
-                    params: {
-                        pipelineId: 123
-                    },
-                    sort: 'descending'
-                })
-                .resolves(getStageMocks(testStages));
-
-            return server.inject(options).then(reply => {
-                assert.equal(reply.statusCode, 200);
-                assert.deepEqual(reply.result, testStages);
-            });
-        });
-
-        it('returns 200 and all stages with matching jobIds', () => {
-            options.url = '/stages?jobIds[]=1&jobIds[]=2&jobIds[]=3&';
-            stageFactoryMock.list
-                .withArgs({
-                    params: {
-                        jobIds: [1, 2, 3]
-                    },
-                    sort: 'descending'
-                })
-                .resolves(getStageMocks(testStages));
-
-            return server.inject(options).then(reply => {
-                assert.equal(reply.statusCode, 200);
-                assert.deepEqual(reply.result, testStages);
-            });
-        });
-
-        it('returns 200 and stages with pagination if no search parameter specified', () => {
-            options.url = '/stages';
-            stageFactoryMock.list.withArgs({ sort: 'descending' }).resolves(getStageMocks(testStages));
-
-            return server.inject(options).then(reply => {
-                assert.equal(reply.statusCode, 200);
-                assert.deepEqual(reply.result, testStages);
-            });
-        });
-
-        it('returns 200 and all stages when sort is set', () => {
-            options.url = '/stages?sort=ascending';
-            stageFactoryMock.list.withArgs({ sort: 'ascending' }).resolves(getStageMocks(testStages));
-
-            return server.inject(options).then(reply => {
-                assert.equal(reply.statusCode, 200);
-                assert.deepEqual(reply.result, testStages);
-            });
-        });
-
-        it('returns 200 and all stages when sortBy is set', () => {
-            options.url = '/stages?sort=ascending&sortBy=name';
-            stageFactoryMock.list
-                .withArgs({
-                    sort: 'ascending',
-                    sortBy: 'name'
-                })
-                .resolves(getStageMocks(testStages));
-
-            return server.inject(options).then(reply => {
-                assert.equal(reply.statusCode, 200);
-                assert.deepEqual(reply.result, testStages);
-            });
-        });
-
-        it('returns 200 and all stages with matched name', () => {
-            options.url = '/stages?page=1&count=3&name=deploy';
-            stageFactoryMock.list
-                .withArgs({
-                    params: {
-                        name: 'deploy'
-                    },
-                    paginate: {
-                        page: 1,
-                        count: 3
-                    },
-                    sort: 'descending'
-                })
-                .resolves(getStageMocks(testStages));
-
-            return server.inject(options).then(reply => {
-                assert.equal(reply.statusCode, 200);
-                assert.deepEqual(reply.result, testStages);
-            });
-        });
-
-        it('returns 500 when datastore fails', () => {
-            stageFactoryMock.list.rejects(new Error('fittoburst'));
-
-            return server.inject(options).then(reply => {
-                assert.equal(reply.statusCode, 500);
-            });
-        });
     });
 
     describe('GET /stages/{id}', () => {
